Scrapes Time-Served as a Representative.
The implementation is nuanced: we must use a very particulah XPath selector which works for Mitch McConell's page but unsure about others. We need to manually extract each of the elements stored in the Information Table on the right panel of the web-page. This is admittedly very tricky and requires hand-crafting regular expressions.

Note that one thing that made scraping time-served particularly tricky was that there was a special character that looked like a hyphen ("-") but was _very_ difficult to recognize using regex, since it was in fact some unicode.